refactor: 세부기능 실행 메소드를 line, Station에서 각각의 Repository로 이동
